dashboard/backend/mqtt-bridge/requirements.txt
paho-mqtt
redis>=5.0.0